Load 3.3.51 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@156 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/strfuns.c
@@ -524,7 +524,7 @@ trim_all_spaces (const string_t * txt)
 
     dest = alloca(mstrsize(txt));
     if (dest == NULL)
-        error("Stack overflow (%ld bytes)\n", mstrsize(txt));
+        error("Stack overflow (%ld bytes)\n", (long)mstrsize(txt));
 
     src = get_txt(txt);
     srclen = mstrsize(txt);
